requirements.txt
alembic
asyncpg
fastapi[all]
passlib[bcrypt]
pydantic-settings
pyjwt
sqlalchemy
uvicorn
httpx
